Change the container name to compile-presentation
The name of the repo was changed in Docker Hub.
The previous name presentation-maker matched the GitHub project
name, but we have a naming style for containers and they should
use a a verb (action) to describe what the container does.

In addition, update the git URL. The project was moved from
the github.com/Aalto-LeTech organization to github.com/apluslms.